// src/enums/generated.ts
export const Button = {
  UP: 0,
  DOWN: 1,
  LEFT: 2,
  RIGHT: 3,
  SUBMIT: 4,
  ACTION: 5,
  CANCEL: 6,
  MENU: 7,
  STATS: 8,
  CYCLE_SHINY: 9,
  CYCLE_FORM: 10,
  CYCLE_GENDER: 11,
  CYCLE_ABILITY: 12,
  CYCLE_NATURE: 13,
  CYCLE_TERA: 14,
  SPEED_UP: 15,
  SLOW_DOWN: 16,
  DEV_CUSTOM: 17,
} as const;
export type Button = (typeof Button)[keyof typeof Button];

export const UiMode = {
  MESSAGE: 0,
  TITLE: 1,
  COMMAND: 2,
  FIGHT: 3,
  BALL: 4,
  TARGET_SELECT: 5,
  MODIFIER_SELECT: 6,
  SAVE_SLOT: 7,
  PARTY: 8,
  SUMMARY: 9,
  STARTER_SELECT: 10,
  EVOLUTION_SCENE: 11,
  EGG_HATCH_SCENE: 12,
  EGG_HATCH_SUMMARY: 13,
  CONFIRM: 14,
  OPTION_SELECT: 15,
  MENU: 16,
  MENU_OPTION_SELECT: 17,
  SETTINGS: 18,
  SETTINGS_DISPLAY: 19,
  SETTINGS_AUDIO: 20,
  SETTINGS_GAMEPAD: 21,
  GAMEPAD_BINDING: 22,
  SETTINGS_KEYBOARD: 23,
  KEYBOARD_BINDING: 24,
  ACHIEVEMENTS: 25,
  GAME_STATS: 26,
  EGG_LIST: 27,
  EGG_GACHA: 28,
  POKEDEX: 29,
  POKEDEX_SCAN: 30,
  POKEDEX_PAGE: 31,
  LOGIN_OR_REGISTER: 32,
  LOGIN_FORM: 33,
  REGISTRATION_FORM: 34,
  LOADING: 35,
  UNAVAILABLE: 36,
  CHALLENGE_SELECT: 37,
  RENAME_POKEMON: 38,
  RENAME_RUN: 39,
  RUN_HISTORY: 40,
  RUN_INFO: 41,
  TEST_DIALOGUE: 42,
  AUTO_COMPLETE: 43,
  ADMIN: 44,
  MYSTERY_ENCOUNTER: 45,
  CHANGE_PASSWORD_FORM: 46,
  ALERT_MODAL: 47,
} as const;
export type UiMode = (typeof UiMode)[keyof typeof UiMode];

export const PartyUiMode = {
  SWITCH: 0,
  FAINT_SWITCH: 1,
  POST_BATTLE_SWITCH: 2,
  REVIVAL_BLESSING: 3,
  MODIFIER: 4,
  MOVE_MODIFIER: 5,
  TM_MODIFIER: 6,
  REMEMBER_MOVE_MODIFIER: 7,
  MODIFIER_TRANSFER: 8,
  SPLICE: 9,
  RELEASE: 10,
  CHECK: 11,
  SELECT: 12,
  DISCARD: 13,
} as const;
export type PartyUiMode = (typeof PartyUiMode)[keyof typeof PartyUiMode];

export const BiomeId = {
  TOWN: 0,
  PLAINS: 1,
  GRASS: 2,
  TALL_GRASS: 3,
  METROPOLIS: 4,
  FOREST: 5,
  SEA: 6,
  SWAMP: 7,
  BEACH: 8,
  LAKE: 9,
  SEABED: 10,
  MOUNTAIN: 11,
  BADLANDS: 12,
  CAVE: 13,
  DESERT: 14,
  ICE_CAVE: 15,
  MEADOW: 16,
  POWER_PLANT: 17,
  VOLCANO: 18,
  GRAVEYARD: 19,
  DOJO: 20,
  FACTORY: 21,
  RUINS: 22,
  WASTELAND: 23,
  ABYSS: 24,
  SPACE: 25,
  CONSTRUCTION_SITE: 26,
  JUNGLE: 27,
  FAIRY_CAVE: 28,
  TEMPLE: 29,
  SLUM: 30,
  SNOWY_FOREST: 31,
  ISLAND: 40,
  LABORATORY: 41,
  END: 50,
} as const;
export type BiomeId = (typeof BiomeId)[keyof typeof BiomeId];

export const StatusEffect = {
  NONE: 0,
  POISON: 1,
  TOXIC: 2,
  PARALYSIS: 3,
  SLEEP: 4,
  FREEZE: 5,
  BURN: 6,
  FAINT: 7,
} as const;
export type StatusEffect = (typeof StatusEffect)[keyof typeof StatusEffect];

export const SaveSlotUiMode = {
  LOAD: 0,
  SAVE: 1,
} as const;
export type SaveSlotUiMode = (typeof SaveSlotUiMode)[keyof typeof SaveSlotUiMode];

export const SummaryUiMode = {
  DEFAULT: 0,
  LEARN_MOVE: 1,
} as const;
export type SummaryUiMode = (typeof SummaryUiMode)[keyof typeof SummaryUiMode];

export const PartyOption = {
  CANCEL: -1,
  SEND_OUT: 0,
  PASS_BATON: 1,
  REVIVE: 2,
  APPLY: 3,
  TEACH: 4,
  TRANSFER: 5,
  SUMMARY: 6,
  POKEDEX: 7,
  UNPAUSE_EVOLUTION: 8,
  SPLICE: 9,
  UNSPLICE: 10,
  RELEASE: 11,
  RENAME: 12,
  SELECT: 13,
  DISCARD: 14,
  SCROLL_UP: 1000,
  SCROLL_DOWN: 1001,
  FORM_CHANGE_ITEM: 2000,
  MOVE_1: 3000,
  MOVE_2: 3001,
  MOVE_3: 3002,
  MOVE_4: 3003,
  ALL: 4000,
} as const;
export type PartyOption = (typeof PartyOption)[keyof typeof PartyOption];

export const AbilityAttr = {
  ABILITY_1: 1,
  ABILITY_2: 2,
  ABILITY_HIDDEN: 4,
} as const;
export type AbilityAttr = (typeof AbilityAttr)[keyof typeof AbilityAttr];
// ...
export type AbilityId = (typeof AbilityId)[keyof typeof AbilityId];

export const AiType = {
  RANDOM: 0,
  SMART_RANDOM: 1,
  SMART: 2,
} as const;
export type AiType = (typeof AiType)[keyof typeof AiType];

export const ArenaTagSide = {
  BOTH: 0,
  PLAYER: 1,
  ENEMY: 2,
} as const;
export type ArenaTagSide = (typeof ArenaTagSide)[keyof typeof ArenaTagSide];

export const BattleType = {
  WILD: 0,
  TRAINER: 1,
  CLEAR: 2,
  MYSTERY_ENCOUNTER: 3,
} as const;
export type BattleType = (typeof BattleType)[keyof typeof BattleType];

export const BattlerIndex = {
  ATTACKER: -1,
  PLAYER: 0,
  PLAYER_2: 1,
  ENEMY: 2,
  ENEMY_2: 3,
} as const;
export type BattlerIndex = (typeof BattlerIndex)[keyof typeof BattlerIndex];

export const BerryType = {
  SITRUS: 0,
  LUM: 1,
  ENIGMA: 2,
  LIECHI: 3,
  GANLON: 4,
  PETAYA: 5,
  APICOT: 6,
  SALAC: 7,
  LANSAT: 8,
  STARF: 9,
  LEPPA: 10,
} as const;
export type BerryType = (typeof BerryType)[keyof typeof BerryType];

export const BiomePoolTier = {
  COMMON: 0,
  UNCOMMON: 1,
  RARE: 2,
  SUPER_RARE: 3,
  ULTRA_RARE: 4,
  BOSS: 5,
  BOSS_RARE: 6,
  BOSS_SUPER_RARE: 7,
  BOSS_ULTRA_RARE: 8,
} as const;
export type BiomePoolTier = (typeof BiomePoolTier)[keyof typeof BiomePoolTier];

export const Challenges = {
  SINGLE_GENERATION: 0,
  SINGLE_TYPE: 1,
  LOWER_MAX_STARTER_COST: 2,
  LOWER_STARTER_POINTS: 3,
  FRESH_START: 4,
  INVERSE_BATTLE: 5,
  FLIP_STAT: 6,
  LIMITED_CATCH: 7,
  LIMITED_SUPPORT: 8,
  HARDCORE: 9,
  PASSIVES: 10,
} as const;
export type Challenges = (typeof Challenges)[keyof typeof Challenges];

export const Command = {
  FIGHT: 0,
  BALL: 1,
  POKEMON: 2,
  RUN: 3,
  TERA: 4,
} as const;
export type Command = (typeof Command)[keyof typeof Command];

export const EvoLevelThresholdKind = {
  STRONG: 0,
  NORMAL: 1,
  WILD: 2,
} as const;
export type EvoLevelThresholdKind = (typeof EvoLevelThresholdKind)[keyof typeof EvoLevelThresholdKind];

export const Gender = {
  GENDERLESS: -1,
  MALE: 0,
  FEMALE: 1,
} as const;
export type Gender = (typeof Gender)[keyof typeof Gender];

export const HitResult = {
  EFFECTIVE: 1,
  EXTREMELY_EFFECTIVE: 2,
  SUPER_EFFECTIVE: 3,
  NOT_VERY_EFFECTIVE: 4,
  MOSTLY_INEFFECTIVE: 5,
  ONE_HIT_KO: 6,
  NO_EFFECT: 7,
  STATUS: 8,
  HEAL: 9,
  FAIL: 10,
  MISS: 11,
  INDIRECT: 12,
  IMMUNE: 13,
  CONFUSION: 14,
  INDIRECT_KO: 15,
} as const;
export type HitResult = (typeof HitResult)[keyof typeof HitResult];

export const ModifierPoolType = {
  PLAYER: 0,
  WILD: 1,
  TRAINER: 2,
  ENEMY_BUFF: 3,
  DAILY_STARTER: 4,
} as const;
export type ModifierPoolType = (typeof ModifierPoolType)[keyof typeof ModifierPoolType];

export const ModifierTier = {
  COMMON: 0,
  GREAT: 1,
  ULTRA: 2,
  ROGUE: 3,
  MASTER: 4,
  LUXURY: 5,
} as const;
export type ModifierTier = (typeof ModifierTier)[keyof typeof ModifierTier];

export const MoveCategory = {
  PHYSICAL: 0,
  SPECIAL: 1,
  STATUS: 2,
} as const;
export type MoveCategory = (typeof MoveCategory)[keyof typeof MoveCategory];

export const MoveFlags = {
  NONE: 0,
  MAKES_CONTACT: 1,
  IGNORE_PROTECT: 2,
  SOUND_BASED: 4,
  HIDE_USER: 8,
  HIDE_TARGET: 16,
  BITING_MOVE: 32,
  PULSE_MOVE: 64,
  PUNCHING_MOVE: 128,
  SLICING_MOVE: 256,
  RECKLESS_MOVE: 512,
  BALLBOMB_MOVE: 1024,
  POWDER_MOVE: 2048,
  DANCE_MOVE: 4096,
  WIND_MOVE: 8192,
  TRIAGE_MOVE: 16384,
  IGNORE_ABILITIES: 32768,
  CHECK_ALL_HITS: 65536,
  IGNORE_SUBSTITUTE: 131072,
  REFLECTABLE: 262144,
  GRAVITY: 524288,
} as const;
export type MoveFlags = (typeof MoveFlags)[keyof typeof MoveFlags];
// ...
export type MoveId = (typeof MoveId)[keyof typeof MoveId];

export const MovePriorityInBracket = {
  LAST: 0,
  NORMAL: 1,
  FIRST: 2,
} as const;
export type MovePriorityInBracket = (typeof MovePriorityInBracket)[keyof typeof MovePriorityInBracket];

export const MoveResult = {
  PENDING: 0,
  SUCCESS: 1,
  FAIL: 2,
  MISS: 3,
  OTHER: 4,
} as const;
export type MoveResult = (typeof MoveResult)[keyof typeof MoveResult];

export const MoveTarget = {
  USER: 0,
  OTHER: 1,
  ALL_OTHERS: 2,
  NEAR_OTHER: 3,
  ALL_NEAR_OTHERS: 4,
  NEAR_ENEMY: 5,
  ALL_NEAR_ENEMIES: 6,
  RANDOM_NEAR_ENEMY: 7,
  ALL_ENEMIES: 8,
  ATTACKER: 9,
  NEAR_ALLY: 10,
  ALLY: 11,
  USER_OR_NEAR_ALLY: 12,
  USER_AND_ALLIES: 13,
  ALL: 14,
  USER_SIDE: 15,
  ENEMY_SIDE: 16,
  BOTH_SIDES: 17,
  PARTY: 18,
  CURSE: 19,
} as const;
export type MoveTarget = (typeof MoveTarget)[keyof typeof MoveTarget];

export const MoveUseMode = {
  NORMAL: 1,
  IGNORE_PP: 2,
  INDIRECT: 3,
  FOLLOW_UP: 4,
  REFLECTED: 5,
  DELAYED_ATTACK: 6,
} as const;
export type MoveUseMode = (typeof MoveUseMode)[keyof typeof MoveUseMode];

export const MultiHitType = {
  TWO: 0,
  TWO_TO_FIVE: 1,
  THREE: 2,
  TEN: 3,
  BEAT_UP: 4,
} as const;
export type MultiHitType = (typeof MultiHitType)[keyof typeof MultiHitType];

export const MysteryEncounterOptionMode = {
  DEFAULT: 0,
  DISABLED_OR_DEFAULT: 1,
  DEFAULT_OR_SPECIAL: 2,
  DISABLED_OR_SPECIAL: 3,
} as const;
export type MysteryEncounterOptionMode = (typeof MysteryEncounterOptionMode)[keyof typeof MysteryEncounterOptionMode];

export const MysteryEncounterTier = {
  COMMON: 66,
  GREAT: 40,
  ULTRA: 19,
  ROGUE: 3,
  MASTER: 0,
} as const;
export type MysteryEncounterTier = (typeof MysteryEncounterTier)[keyof typeof MysteryEncounterTier];

export const MysteryEncounterType = {
  MYSTERIOUS_CHALLENGERS: 0,
  MYSTERIOUS_CHEST: 1,
  DARK_DEAL: 2,
  FIGHT_OR_FLIGHT: 3,
  SLUMBERING_SNORLAX: 4,
  TRAINING_SESSION: 5,
  DEPARTMENT_STORE_SALE: 6,
  SHADY_VITAMIN_DEALER: 7,
  FIELD_TRIP: 8,
  SAFARI_ZONE: 9,
  LOST_AT_SEA: 10,
  FIERY_FALLOUT: 11,
  THE_STRONG_STUFF: 12,
  THE_POKEMON_SALESMAN: 13,
  AN_OFFER_YOU_CANT_REFUSE: 14,
  DELIBIRDY: 15,
  ABSOLUTE_AVARICE: 16,
  A_TRAINERS_TEST: 17,
  TRASH_TO_TREASURE: 18,
  BERRIES_ABOUND: 19,
  CLOWNING_AROUND: 20,
  PART_TIMER: 21,
  DANCING_LESSONS: 22,
  WEIRD_DREAM: 23,
  THE_WINSTRATE_CHALLENGE: 24,
  TELEPORTING_HIJINKS: 25,
  BUG_TYPE_SUPERFAN: 26,
  FUN_AND_GAMES: 27,
  UNCOMMON_BREED: 28,
  GLOBAL_TRADE_SYSTEM: 29,
  THE_EXPERT_POKEMON_BREEDER: 30,
} as const;
export type MysteryEncounterType = (typeof MysteryEncounterType)[keyof typeof MysteryEncounterType];

export const Nature = {
  HARDY: 0,
  LONELY: 1,
  BRAVE: 2,
  ADAMANT: 3,
  NAUGHTY: 4,
  BOLD: 5,
  DOCILE: 6,
  RELAXED: 7,
  IMPISH: 8,
  LAX: 9,
  TIMID: 10,
  HASTY: 11,
  SERIOUS: 12,
  JOLLY: 13,
  NAIVE: 14,
  MODEST: 15,
  MILD: 16,
  QUIET: 17,
  BASHFUL: 18,
  RASH: 19,
  CALM: 20,
  GENTLE: 21,
  SASSY: 22,
  CAREFUL: 23,
  QUIRKY: 24,
} as const;
export type Nature = (typeof Nature)[keyof typeof Nature];

export const Passive = {
  UNLOCKED: 1,
  ENABLED: 2,
} as const;
export type Passive = (typeof Passive)[keyof typeof Passive];

export const PokeballType = {
  POKEBALL: 0,
  GREAT_BALL: 1,
  ULTRA_BALL: 2,
  ROGUE_BALL: 3,
  MASTER_BALL: 4,
  LUXURY_BALL: 5,
} as const;
export type PokeballType = (typeof PokeballType)[keyof typeof PokeballType];
// ...
export const NAMES = {
  Button: invert(Button),
  UiMode: invert(UiMode),
  PartyUiMode: invert(PartyUiMode),
  BiomeId: invert(BiomeId),
  StatusEffect: invert(StatusEffect),
  SaveSlotUiMode: invert(SaveSlotUiMode),
  SummaryUiMode: invert(SummaryUiMode),
  PartyOption: invert(PartyOption),
  AbilityAttr: invert(AbilityAttr),
  AbilityId: invert(AbilityId),
  AiType: invert(AiType),
  ArenaTagSide: invert(ArenaTagSide),
  BattleType: invert(BattleType),
  BattlerIndex: invert(BattlerIndex),
  BerryType: invert(BerryType),
  BiomePoolTier: invert(BiomePoolTier),
  Challenges: invert(Challenges),
  Command: invert(Command),
  EvoLevelThresholdKind: invert(EvoLevelThresholdKind),
  Gender: invert(Gender),
  HitResult: invert(HitResult),
  ModifierPoolType: invert(ModifierPoolType),
  ModifierTier: invert(ModifierTier),
  MoveCategory: invert(MoveCategory),
  MoveFlags: invert(MoveFlags),
  MoveId: invert(MoveId),
  MovePriorityInBracket: invert(MovePriorityInBracket),
  MoveResult: invert(MoveResult),
  MoveTarget: invert(MoveTarget),
  MoveUseMode: invert(MoveUseMode),
  MultiHitType: invert(MultiHitType),
  MysteryEncounterOptionMode: invert(MysteryEncounterOptionMode),
  MysteryEncounterTier: invert(MysteryEncounterTier),
  MysteryEncounterType: invert(MysteryEncounterType),
  Nature: invert(Nature),
  Passive: invert(Passive),
  PokeballType: invert(PokeballType),
  PokemonType: invert(PokemonType),
  SpeciesId: invert(SpeciesId),
  Stat: invert(Stat),
  TerrainType: invert(TerrainType),
  TimeOfDay: invert(TimeOfDay),
  TrainerPoolTier: invert(TrainerPoolTier),
  TrainerSlot: invert(TrainerSlot),
  WeatherType: invert(WeatherType),
} as const;